[refactor/#29] 코드리뷰 반영 (이름 편경)
