+++ app/src/main/java/sopt/org/starbucks/core/navigation/Route.kt
@@ -23,6 +23,3 @@ data object Other : MainTabRoute
 
 @Serializable
 data object MyMenu : Route
-
-@Serializable
-data object EditMenu : Route
