+++ app/src/main/java/sopt/org/starbucks/ui/main/MainNavHost.kt
@@ -13,7 +13,6 @@ import sopt.org.starbucks.core.navigation.Other
 import sopt.org.starbucks.core.navigation.Pay
 import sopt.org.starbucks.core.navigation.Shop
 import sopt.org.starbucks.ui.home.HomeRoute
-import sopt.org.starbucks.ui.mymenu.MyMenuRoute
 import sopt.org.starbucks.ui.order.OrderRoute
 
 @Composable
@@ -42,9 +41,9 @@ fun MainNavHost(
         composable<Shop> { }
         composable<Other> { }
         composable<MyMenu> {
-            MyMenuRoute(
-                paddingValues = paddingValues
-            )
+//            MyMenuRoute(
+//                paddingValues = paddingValues
+//            )
         }
     }
 }
