+++ app/src/main/java/sopt/org/starbucks/ui/mymenu/MyMenuUiState.kt
@@ -4,7 +4,7 @@ import sopt.org.starbucks.data.model.MenuDetailModel
 import sopt.org.starbucks.ui.mymenu.component.DrinkSize
 import sopt.org.starbucks.ui.mymenu.component.TabType
 
-data class EditMenuUiState(
+data class MyMenuUiState(
     val menu: MenuDetailModel = MenuDetailModel(),
     val selectedTab: TabType = TabType.ICED,
     val selectedSize: DrinkSize = DrinkSize.TALL,
